Move report_cell_and_block_counts() from vlasiator.cpp to memory_report.cpp.
