+++ vlasiator.cpp
@@ -92,45 +92,6 @@ void addTimedBarrier(string name){
    MPI_Barrier(MPI_COMM_WORLD);
 }
 
-
-/*! Report spatial cell counts per refinement level as well as velocity cell counts per population into logfile
- */
-void report_cell_and_block_counts(dccrg::Dccrg<spatial_cell::SpatialCell,dccrg::Cartesian_Geometry>& mpiGrid){
-   cint maxRefLevel = mpiGrid.get_maximum_refinement_level();
-   const vector<CellID> localCells = getLocalCells();
-   cint popCount = getObjectWrapper().particleSpecies.size();
-
-   // popCount+1 as we store the spatial cell counts and then the populations' v_cell counts.
-   // maxRefLevel+1 as e.g. there's 2 levels at maxRefLevel == 1
-   std::vector<int64_t> localCounts((popCount+1)*(maxRefLevel+1), 0), globalCounts((popCount+1)*(maxRefLevel+1), 0);
-
-   for (const auto cellid : localCells) {
-      cint level = mpiGrid.get_refinement_level(cellid);
-      localCounts[level]++;
-      for(int pop=0; pop<popCount; pop++) {
-         localCounts[maxRefLevel+1 + level*popCount + pop] += mpiGrid[cellid]->get_number_of_velocity_blocks(pop);
-      }
-   }
-
-   MPI_Reduce(localCounts.data(), globalCounts.data(), (popCount+1)*(maxRefLevel+1), MPI_INT64_T, MPI_SUM, MASTER_RANK, MPI_COMM_WORLD);
-
-   logFile << "(CELLS) tstep = " << P::tstep << " time = " << P::t << " spatial cells [ ";
-   for(int level = 0; level <= maxRefLevel; level++) {
-      logFile << globalCounts[level] << " ";
-   }
-   logFile << "] blocks ";
-   for(int pop=0; pop<popCount; pop++) {
-      logFile << getObjectWrapper().particleSpecies[pop].name << " [ ";
-      for(int level = 0; level <= maxRefLevel; level++) {
-         logFile << globalCounts[maxRefLevel+1 + level*popCount + pop] << " ";
-      }
-      logFile << "] ";
-   }
-   logFile << endl << flush;
-
-}
-
-
 void computeNewTimeStep(dccrg::Dccrg<SpatialCell,dccrg::Cartesian_Geometry>& mpiGrid,
 			FsGrid< fsgrids::technical, FS_STENCIL_WIDTH> & technicalGrid, Real &newDt, bool &isChanged) {
